% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_matrix.R
\name{as.sgadj}
\alias{as.sgadj}
\title{Creator for sgadj-class}
\usage{
as.sgadj(edges = NULL, type = "?", pars = NULL, other = "")
}
\arguments{
\item{edges}{edge list-of-lists}

\item{type}{of the graph}

\item{pars}{parameters for the graph}

\item{other}{other comments}
}
\description{
Creator for sgadj-class
}
