% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/class_sgc.R
\name{as.sgc}
\alias{as.sgc}
\title{Creator for sgc}
\usage{
as.sgc(clusters, type = "?", pars = NULL, note = NULL)
}
\arguments{
\item{clusters}{list of clusters as point indices}

\item{type}{type}

\item{pars}{parameters}

\item{note}{notes}
}
\description{
Creator for sgc
}

