% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cut_sg.R
\name{cut.sg}
\alias{cut.sg}
\title{cut edges}
\usage{
\method{cut}{sg}(x, data, R, ...)
}
\arguments{
\item{x}{sg graph object}

\item{data}{point pattern used for computing g}

\item{R}{cutting length}

\item{...}{ignored

Removes edges with length > R.}
}
\description{
cut edges
}

