% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_sg.R
\name{plot.sg}
\alias{plot.sg}
\title{Plot a spatial graph}
\usage{
\method{plot}{sg}(x, data, which = NULL, add = FALSE, addPoints = FALSE,
  points.pch = 1, points.col = 1, points.cex = 1, max.edges = 10000,
  ...)
}
\arguments{
\item{x}{an 'sg' graph object}

\item{data}{The point pattern object, same as for computing the 'g'}

\item{which}{Indices of which out-edges to plot. Default: all}

\item{add}{Add to existing plot? (default: FALSE)}

\item{addPoints}{Add points? Will be added if add=FALSE}

\item{points.pch}{point styling}

\item{points.col}{point styling}

\item{points.cex}{point styling}

\item{max.edges}{limit of edges to try to plot, gets very slow at high count. default 1e4}

\item{...}{passed to 'lines' function}
}
\description{
Rudimentary plotting.
}

