% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/adjacency_matrix.R
\name{plot.sgadj}
\alias{plot.sgadj}
\title{plot sgadj}
\usage{
\method{plot}{sgadj}(x, ...)
}
\arguments{
\item{x}{sgadj object}

\item{...}{passed to plot.sg

converts to sg and plots that.}
}
\description{
plot sgadj
}

