% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spectral_sg.R
\name{plot.sgspectral}
\alias{plot.sgspectral}
\title{plot spectral clustering results}
\usage{
\method{plot}{sgspectral}(x, data, ...)
}
\arguments{
\item{x}{spectral.sg result}

\item{data}{point pattern}

\item{...}{ignored}
}
\description{
plot spectral clustering results
}

