% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sg.R
\name{plot3_sg}
\alias{plot3_sg}
\title{Plot 3d graph}
\usage{
plot3_sg(x, data, which, ...)
}
\arguments{
\item{x}{sg object}

\item{data}{coordinates}

\item{which}{points of which out-edges will be plotted}

\item{...}{passed to rgl.lines}
}
\description{
Plot 3d graph
}
