% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sg2.R
\name{print.sg}
\alias{print.sg}
\title{Print method for sg}
\usage{
\method{print}{sg}(x, ...)
}
\arguments{
\item{x}{sg object}

\item{...}{ignored}
}
\description{
Print sg class.
}
\details{
Print basic info.
}
