% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/prune_sg.R
\name{prune.sg}
\alias{prune.sg}
\title{Prune a graph}
\usage{
prune.sg(g, level = 1, verbose = FALSE)
}
\arguments{
\item{g}{sg object}

\item{level}{pruning level}

\item{verbose}{verbosity}
}
\description{
Prune a graph
}

