% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune_sg.R
\name{prune_sg}
\alias{prune_sg}
\title{Prune a graph}
\usage{
prune_sg(g, level = 1, verbose = FALSE)
}
\arguments{
\item{g}{sg object}

\item{level}{pruning level}

\item{verbose}{verbosity}
}
\description{
Prune a graph
}
\details{
Remove edges from a graph by their path connectivity.
}
\examples{
x <- matrix(runif(50*2), ncol=2)
g <- spatgraph(x, "MST")
gp <- prune_sg(g, level = 2)
plot(g, x, lty=2)
plot(gp, x, add=TRUE, col=2)

}
