% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg2dxf.R
\name{sg2dxf}
\alias{sg2dxf}
\title{sg to dxf format}
\usage{
sg2dxf(g, x, file)
}
\arguments{
\item{g}{sg object}

\item{x}{pattern object used for computing g}

\item{file}{filename for output}
}
\description{
sg to dxf format
}
