% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg2igraph.R
\name{sg2igraph}
\alias{sg2igraph}
\title{sg to igraph}
\usage{
sg2igraph(g, x, ...)
}
\arguments{
\item{g}{sg object}

\item{x}{possibly the location pattern}

\item{...}{not used}
}
\description{
sg to igraph
}
