% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_matrix.R
\name{sg2sparse}
\alias{sg2sparse}
\title{Make a sparse adjacency matrix from sg-object}
\usage{
sg2sparse(x)
}
\arguments{
\item{x}{sg-object}
}
\description{
Make a sparse adjacency matrix from sg-object
}
