% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constants.R
\docType{data}
\name{SG_GRAPH_PARAMETERS}
\alias{SG_GRAPH_PARAMETERS}
\title{Supported graphs constants}
\format{\preformatted{List of 10
 $ geometric     :List of 1
  ..$ R: chr "numeric>0"
 $ knn           :List of 1
  ..$ k: chr "integer>0"
 $ mass_geometric:List of 1
  ..$ mass: chr "numeric vector of sizes"
 $ markcross     :List of 1
  ..$ mass: chr "numeric vector of sizes"
 $ gabriel       : list()
 $ MST           : list()
 $ SIG           : list()
 $ RST           :List of 1
  ..$ center: chr "coordinates of the center"
 $ RNG           : list()
 $ CCC           :List of 1
  ..$ types: chr "factor vector of types"
}}
\usage{
SG_GRAPH_PARAMETERS
}
\description{
Supported graphs constants
}
\keyword{datasets}

