% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/parse_input_coordinates.R
\name{sg_parse_coordinates}
\alias{sg_parse_coordinates}
\title{Parse input for coordinates}
\usage{
sg_parse_coordinates(x, verbose = FALSE)
}
\arguments{
\item{x}{Input object containing the coordinates in some format.}

\item{verbose}{Print out info of the coordinates.}
}
\description{
Extract the coordinate locations from the input object.
}

