% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/verify_parameters.R
\name{sg_verify_parameters}
\alias{sg_verify_parameters}
\title{Verify input parameters for the graph}
\usage{
sg_verify_parameters(coord, type, par, maxR, doDists, preGraph)
}
\arguments{
\item{coord}{Coordinates of the locations}

\item{type}{Type of graph}

\item{par}{Parameter(s) for the graph}

\item{maxR}{Maximum range for edges, helps in large patterns.}

\item{doDists}{Precompute distances? Speeds up some graphs, takes up memory.}

\item{preGraph}{Precomputed graph, taken as a super-graph}
}
\description{
Mainly for internal use.
}
