% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/shortestPath.R
\name{shortestPath}
\alias{shortestPath}
\title{shortest path on the graph}
\usage{
shortestPath(i, j, g, x = NULL, dbg = FALSE)
}
\arguments{
\item{i}{index from}

\item{j}{index to}

\item{g}{sg object}

\item{x}{optional point pattern from which g was computed}

\item{dbg}{verbose}
}
\description{
Djikstra's algorithm
}

