% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortestPath_c.R
\name{shortestPath2}
\alias{shortestPath2}
\title{shortest path on the graph}
\usage{
shortestPath2(i, j, g, x = NULL, dbg = FALSE, checksym = TRUE)
}
\arguments{
\item{i}{index from}

\item{j}{index to. Can be a set.}

\item{g}{sg object}

\item{x}{optional point pattern from which g was computed}

\item{dbg}{verbose}

\item{checksym}{check (and force) symmetry}
}
\description{
Djikstra's algorithm
}
\details{
If x is given, we use the point-to-point distances as edge weights. Otherwise, each
edge has weight 1.
}
