% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shortestPath_old.R
\name{shortestPath_legacy}
\alias{shortestPath_legacy}
\title{shortest path on the graph}
\usage{
shortestPath_legacy(i, j, g, x = NULL, dbg = FALSE)
}
\arguments{
\item{i}{index from}

\item{j}{index to}

\item{g}{sg object}

\item{x}{optional point pattern from which g was computed}

\item{dbg}{verbose}
}
\description{
Djikstra's algorithm
}
\details{
If x is given, we use the point-to-point distances as edge weights. Otherwise, each
edge has weight 1.
}
