% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_matrix.R
\name{sparse2sg}
\alias{sparse2sg}
\title{Make an sg-object from adjacency matrix}
\usage{
sparse2sg(x)
}
\arguments{
\item{x}{square matrix. non-0 elements are taken as edge presence.}
}
\description{
Make an sg-object from adjacency matrix
}
