% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/spatcluster.R
\name{spatcluster}
\alias{spatcluster}
\title{Compute the connected components of a graph}
\usage{
spatcluster(x, verbose = TRUE, sym = FALSE)
}
\arguments{
\item{x}{sg-object}

\item{verbose}{print info}

\item{sym}{force symmetry of edges}
}
\description{
Compute the connected components of a graph
}

