\name{spatgraphs-clusfun}
\alias{clusfun}
\title{clusfun}
\description{
Calculate for a given point pattern the clustering coefficient functional as a function of geometric graph with parameter r 
}

\details{
  \tabular{ll}{
    Date: \tab 2007-10-16\cr
    License: \tab GPL v2 or later\cr
}
}

\arguments{
\item{pp}{The point pattern under study.}
\item{rvec}{Values for which the geometric graph and further the clustering coefficient is to be calculated.}
}

\references{
Watts, D.J. and Strogatz, S.: Collective dynamics of \'small-world\' networks. Nature 393: 440-442, 1998.
}
\keyword{spatial}
