\name{spatgraphs-cumconfun}
\alias{cumconfun}
\title{cumconfun}
\description{
  Cumulative version of connectivity function (see confun).
}
\details{
  \tabular{ll}{
    Date: \tab 2007-10-16\cr
    License: \tab GPL v2 or later\cr
  }
}

\arguments{
\item{pp}{Point pattern.}
\item{rvec}{Distance values of interest.}
\item{type}{Passed to spatgraph().}
\item{pars}{Passed to spatgraph().}
\item{toroidal.correction}{Toroidal correction.}
}
\references{
Illian, Penttinen, Stoyan, Stoyan: Statistical analysis and modelling of spatial point patterns. Springer, 2008.
}
\keyword{spatial}