\name{spatgraphs-cut.sg}
\alias{cut.sg}
\alias{cut_sg}
\alias{prune.sg}
\alias{prune_sg}
\alias{prune}
\title{cut and prune}

\description{
  Cut and prune MST (or others).
  
  Cut the edges of graph with length more that R>0.
  
  Prune 'cuts' away all branches of the graph which are shorter than level>0.
}
\details{
  \tabular{ll}{
    Date: \tab 2008-07-30\cr
    License: \tab GPL v2 or later\cr
  }
}
\keyword{spatial}