\name{spatgraphs-adj2sg}
\alias{adj2sg}
\alias{sg2adj}
\alias{sgadj}
\alias{sg2sparse}
\alias{sparse2sg}


\title{Adjacency matrix to edgelist and vice versa}

\description{
Maps between edgelist and the adjancency matrix representation. Works also with sparse matrices using Matrix-package.
  
}
\usage{
   sg2adj(x)
   adj2sg(x)
   sg2sparse(x)
   sparse2sg(x)
}
\arguments{
\item{x}{sg-object, sgadj-object, or sparseMatrix-object.}
}

\keyword{spatial}