\name{spatgraphs-mailer}
\alias{mailer}
\alias{took}
\title{mailer}

\description{
  Mailer: Unix/Linux only, needs the program 'mail': send an e-mail. Useful to notify ending of long calculation on remote computers.
  
  Took: Nice formatting of time lapsed since given Sys.time-object.
}
\details{
  \tabular{ll}{
    Date: \tab 2008-07-30\cr
    License: \tab GPL v2 or later\cr
  }
}
\keyword{spatial}