\name{spatgraphs-mst_prune}
\alias{mst_prune}
\title{mst\_prune}
\description{
  Prune the MST-graph.
}
\details{
  \tabular{ll}{
    Date: \tab 2007-10-16\cr
    License: \tab GPL v2 or later\cr
  }
}

\arguments{
\item{mst}{Adjacency matrix of the MST. If none given, use pp to compute it.}
\item{pp}{Point pattern. Not needed if mst is provided.}
\item{level}{At which level to prune the MST.}

}
\keyword{spatial}