\name{spatgraphs-plot.sg}
\alias{plot.sg}
\alias{plot_sg}
\alias{plot_sgc}
\alias{plot.sgc}
\alias{plot.sgadj}
\title{plot.sg}
\description{
  Plot the edges of graph, or color the clusters.
}
\details{
  \tabular{ll}{
    Date: \tab 2008-07-30\cr
    License: \tab GPL v2 or later\cr
  }
}

\arguments{
\item{x}{spatgraph/spatcluster object}
\item{pp}{point pattern}
\item{...}{color, linesize etc. for corresponding function (lines, points, points3d).}
}
\keyword{spatial}