\name{spatgraphs-plot_clusters}
\alias{plot_clusters}
\title{plot\_clusters}
\description{
Color the points according to the clusters.
}
\details{
  \tabular{ll}{
    Date: \tab 2007-10-16\cr
    License: \tab GPL v2 or later\cr
  }
}

\arguments{
\item{pp}{Point pattern.}
\item{ec}{Cluster adjacency matrix, from sgcluster.}
\item{pointsize}{Size of points.}
\item{add}{If to only add or to create new plot.}
\item{b}{Starting index of colors().}
}
\keyword{spatial}