\name{spatgraphs-print.sg}
\alias{print.sg}
\alias{print.sgadj}
\alias{print.sgc}
\title{print.sg}
\description{
  Print method of sg, sgadj and sgc-object of package spatgraphs.
}
\details{
  \tabular{ll}{
    Date: \tab 2007-10-24\cr
    License: \tab GPL v2 or later\cr
  }
}

\keyword{spatial}


