\name{spatgraphs-rclustermatrix}
\alias{rclustermatrix}
\title{rclustermatrix}
\description{
  Compute the clustermatrix from adjacency matrix. This is a n x n symmetric square matrix with [i,j]=1 indicating that point x\_i is in the same component as x\_j.
}

\details{
  \tabular{ll}{
    Date: \tab 2007-03-01\cr
    License: \tab GPL v2 or later\cr
}
}

\arguments{
\item{e}{Symmetric adjacency matrix}
}

\value{
The clustermatrix.
}
\keyword{spatial}
