\name{spatgraphs-rcross_graph}
\alias{rcross_graph}
\title{rcross\_graph}
\description{
  Compute the adjacency matrix for marked point pattern with relation
\preformatted{
    x~y <==> ||x-y||<m(x)+m(y)
}  

}

\details{
  \tabular{ll}{
    Date: \tab 2007-03-01\cr
    License: \tab GPL v2 or later\cr
}
}

\arguments{
\item{pp}{Marked point pattern with elements x,y,mark,n. }
}

\keyword{spatial}
