\name{spatgraphs-rgeo_graph}
\alias{rgeo_graph}
\title{rgeo\_graph}
\description{
  Compute the adjacency matrix for point pattern X with relation
\preformatted{
    x~y <==> ||x-y||<R
}

}

\details{
  \tabular{ll}{
    Date: \tab 2007-03-01\cr
    License: \tab GPL v2 or later\cr
}
}

\arguments{
\item{pp}{Point pattern with elements x,y,n. }
\item{R}{The distance with which to connect points}
}

\keyword{spatial}
