\name{spatgraphs-rlist_clusters}
\alias{rlist_clusters}
\title{rlist\_clusters}
\description{
  Give the clustermatrix information as a list of cluster members.
}

\details{
  \tabular{ll}{
    Date: \tab 2007-03-01\cr
    License: \tab GPL v2 or later\cr
}
}

\arguments{
\item{ec}{Clustermatrix}
}

\keyword{spatial}
