\name{spatgraphs-rmark_graph}
\alias{rmark_graph}
\title{rmark\_graph}
\description{
  Compute the adjacency matrix for marked point pattern X with relation
    \preformatted{
  x~y <==> ||x-y||<m(x)
}

}

\details{
  \tabular{ll}{
    Date: \tab 2007-03-01\cr
    License: \tab GPL v2 or later\cr
}
}

\arguments{
\item{pp}{Point pattern with members x,y,n}
}

\keyword{spatial}
