\name{spatgraphs-rmst_graph}
\alias{rmst_graph}
\title{rmst\_graph}
\description{
  Compute the adjacency matrix for point pattern X using the minimum spanning tree rule. That is, compute the spanning tree of the point pattern such that the sum of the resulting edges is minimal. Computed using Prim's algorithm. See e.g. \preformatted{http://en.wikipedia.org/wiki/Prim's_algorithm} .
}

\details{
  \tabular{ll}{
    Date: \tab 2007-03-01\cr
    License: \tab GPL v2 or later\cr
}
}

\arguments{
\item{pp}{Point pattern with (at least) members x,y,n}
}

\keyword{spatial}
