\name{spatgraphs-rnn_graph}
\alias{rnn_graph}
\title{rnn\_graph}
\description{
  Compute the adjacency matrix for point pattern X using the relation
\preformatted{
    x~y <==> ||x-y||<= min_z ||x-z||
}
}

\details{
  \tabular{ll}{
    Date: \tab 2007-03-01\cr
    License: \tab GPL v2 or later\cr
}
}

\arguments{
\item{pp}{Point pattern with (at least) members x,y,n}
}

\keyword{spatial}
