\name{spatgraphs-runif3d}
\alias{runif3d}
\alias{spin3d}
\title{runif3d and spin3d}

\description{
  Simple simulation of uniform 3d point pattern.
  
  spin3d simply spins the rgl-scenery.
}
\details{
  \tabular{ll}{
    Date: \tab 2008-07-30\cr
    License: \tab GPL v2 or later\cr
  }
}
\keyword{spatial}