\name{spatgraphs-sg}
\alias{sg}
\alias{verifyclass}
\alias{summary.sg}
\alias{t.sg}
\alias{t.sgadj}
\alias{summary.sgc}
\title{sg}


\description{
  Edge list-of-lists class for spatgraphs. Methods: print, plot, summary, t.

  Transposing is reversing edges.
}
\details{
  \tabular{ll}{
    Date: \tab 2010-09-29\cr
    License: \tab GPL v2 or later\cr
  }
}
\keyword{spatial}