\name{spatgraphs-sg2dxf}
\alias{sg2dxf}
\title{sg2dxf}
\description{
  Write the graph to an AutoCAD Drawing Exchange Format or dxf file.
}
\details{
  \tabular{ll}{
    Date: \tab 2008-07-30\cr
    License: \tab GPL v2 or later\cr
  }
}
\arguments{
\item{x}{spatgraph object.}
\item{pp}{point pattern.}
\item{file}{output filename.}
}

\keyword{spatial}