\name{spatgraphs-sgcluster}
\alias{sgcluster}
\title{sgcluster}
\description{
Given a adjacency matrix, figure out which points are in the same connected component. Return a nxn-matrix with indicators for such an events.
}

\details{
  \tabular{ll}{
    Date: \tab 2007-10-16\cr
    License: \tab GPL v2 or later\cr
}
}

\arguments{
\item{e}{The adjacency matrix.}
\item{print_dbg}{Print additional information.}
}

\keyword{spatial}
