\name{spatgraphs-sgclusterlist}
\alias{sgclusterlist}
\title{sgclusterlist}
\description{
  Produce a list of groups according to the clusters of the graph.
}

\details{
  \tabular{ll}{
    Date: \tab 2007-10-16\cr
    License: \tab GPL v2 or later\cr
}
}

\arguments{
\item{ec}{Cluster adjacency matrix given by sgcluster.}
}

\keyword{spatial}
