\name{spatgraph-shake}
\alias{shake}
\title{shake}

\description{
  Shake (displace) the points a little. Grid like sampling gives problems with e.g. Delauney graph (collinearity).
  Similar to \code{jitter} in \code{spatstat}
}
\details{
  \tabular{ll}{
    Date: \tab 2008-07-30\cr
    License: \tab GPL v2 or later\cr
  }
}
\keyword{spatial}