\name{spatgraphs-spatcluster}
\alias{spatcluster}
\title{spatcluster}
\description{
  Compute the list of clusters i.e. connected components given a graph-object from spatgraph().
}
\details{
  \tabular{ll}{
    Date: \tab 2008-07-29\cr
    License: \tab GPL v2 or later\cr
  }
}
\arguments{
\item{x}{spatgraph-object}
\item{dbg}{Boolean, print additional messages}
\item{sym}{Boolean, symmetricise the graph first?}
}
\keyword{spatial}