\name{spatgraphs-spin3d}
\alias{spin3d}
\title{spin3d}
\description{
  Spin a rgl scenery for a give time period. Just for visual delight. See "rgl" for more info.
}
\details{
  \tabular{ll}{
    Date: \tab 2007-10-16\cr
    License: \tab GPL v2 or later\cr
  }
}

\arguments{
\item{sec}{Time to spin in seconds.}
\item{rotspeed}{Speed of rotation.}
\item{ang}{At which angle.}
\item{zoom}{Zoom factor.}
\item{fov}{Field of view.}
}
\keyword{spatial}