% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectral_sg.R
\name{spectral.sg}
\alias{spectral.sg}
\title{spectral clustering}
\usage{
spectral.sg(g, m = 2, K = 3)
}
\arguments{
\item{g}{sg object. Should be weighted (with weight.sg-function)}

\item{m}{levels to consider}

\item{K}{number of assumed clusters}
}
\description{
spectral clustering
}
