% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_matrix.R
\name{t.sg}
\alias{t.sg}
\title{Transpose sg object}
\usage{
\method{t}{sg}(x)
}
\arguments{
\item{x}{sg-object.}
}
\description{
This will transpose the adjacency matrix underlying the graph. Will transform
to and from sgadj-object (see 'sg2adj')
}
