% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_matrix.R
\name{t.sgadj}
\alias{t.sgadj}
\title{Transpose sgadj object}
\usage{
\method{t}{sgadj}(x)
}
\arguments{
\item{x}{sgadj object}
}
\description{
This will transpose the adjacency matrix underlying the graph.
}
