% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/set_weights.R
\name{weight.sg}
\alias{weight.sg}
\title{Set weights to edges of sg}
\usage{
weight.sg(g, x, f = function(x) exp(-x^2/scale), scale = 1, ...)
}
\arguments{
\item{g}{sg object}

\item{x}{point pattern used in g}

\item{f}{function for the weight}

\item{scale}{additional scale parameter for the default f}

\item{...}{ignored}
}
\description{
For each edge e(i,j) between points i,j, set the weight f(||x_i-x_j||)
}
\details{
Default f(x) = exp(-x^2/scale)
}

