% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_sg2.R
\name{as.sg}
\alias{as.sg}
\title{Class creator}
\usage{
as.sg(edges = list(), type = "?", pars = NULL, note = NULL)
}
\arguments{
\item{edges}{list of neighbourhoods}

\item{type}{type}

\item{pars}{parameters}

\item{note}{notes}
}
\description{
Class creator
}
