% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_sgc.R
\name{plot.sgc}
\alias{plot.sgc}
\title{plot clusters}
\usage{
\method{plot}{sgc}(x, data, atleast = 2, add = FALSE, col, ...)
}
\arguments{
\item{x}{spatcluster-cluster object}

\item{data}{point pattern object used for computing the graph}

\item{atleast}{plot only cluster with 'atleast' points in them}

\item{add}{add or plot new}

\item{col}{colors for clusters, chosen randomly if missing.}

\item{...}{passed to points}
}
\description{
plot clusters
}

