% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjacency_matrix.R
\name{print.sgadj}
\alias{print.sgadj}
\title{print method for sgadj}
\usage{
\method{print}{sgadj}(x, ...)
}
\arguments{
\item{x}{sgadj object}

\item{...}{ignored}
}
\description{
print method for sgadj
}
