% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sg2igraph.R
\name{sg2igraph}
\alias{sg2igraph}
\title{sg to igraph}
\usage{
sg2igraph(g, x, ...)
}
\arguments{
\item{g}{sg object}

\item{x}{possibly the location pattern}

\item{...}{not used

Not implemented. You can use the 'graph_from_adj_list'-function in 'igraph'-package on the edges-element of the graph.}
}
\description{
Obsolete. Use igraph::graph_from_adj_list on the graph edges element.
}
\examples{
\dontrun{
ix <- igraph::graph_from_adj_list(x$edges)
}

}
