\name{spatgraphs-adj2sg}
\alias{adj2sg}
\alias{sg2adj}
\alias{sgadj}

\title{Adjacency matrix to edgelist and vice versa}

\description{
Maps between edgelist and the adjancency matrix representation.  
}
\details{
  \tabular{ll}{
    Date: \tab 2008-07-30\cr
    License: \tab GPL v2 or later\cr
  }
}
\usage{
   sg2adj(x)
   adj2sg(x)
}
\arguments{
\item{x}{sg-object or sgadj-object.}
}

\keyword{spatial}