\name{spatgraphs-plot_clusters}
\alias{plot_clusters}
\title{plot\_clusters}

\description{
  Plot the clusters using information in the clustermatrix.
}

\details{
  \tabular{ll}{
    Date: \tab 2007-03-01\cr
    License: \tab GPL v2 or later\cr
  }
}

\arguments{
\item{pp} {Point pattern with at least members x,y,n}
\item{ec} {Clustermatrix from rclustermatrix}
\item{b=2}{Starting color, index of colors()}
\item{d=1}{Point size}
}


\keyword{spatial}
