\name{spatgraphs-plot_graph}
\alias{plot_graph}
\title{plot\_graph}
\description{
  Plot the graph on top of point pattern plot using the information of the neighbourhood matrix.
}

\details{
  \tabular{ll}{
    Date: \tab 2007-03-01\cr
    License: \tab GPL v2 or later\cr
}
}

\arguments{
\item{pp}{Point pattern with at least members x,y,n}
\item{edges}{Neighbourhood matrix}
\item{col=1}{Colour}
\item{lty=1}{Line type, 1=solid,2=dashed etc.}
}

\keyword{spatial}
