\name{spatgraphs-runif3d}
\alias{runif3d}
\title{runif3d}
\description{
  Simple uniform point pattern simulation in a 3D-box.
}
\details{
  \tabular{ll}{
    Date: \tab 2007-10-24\cr
    License: \tab GPL v2 or later\cr
  }
}
\arguments{
{n}{Number of points to generate}
{window}{List with members for x,y and z limits, e.g x=c(0,1).}
}

\keyword{spatial}


