\name{spatgraphs-sg}
\alias{sg}
\alias{sgadj}
\alias{sg2adj}
\alias{adj2sg}
\alias{t.sg}
\alias{t.sgadj}
\alias{verifyclass}
\alias{summary.sg}
\alias{summary.sgc}
\title{sg}


\description{
  Edge list-of-lists class for spatgraphs. Methods: print, plot.
  
  Older version used adjacency matrix, this class is now called sgadj.
  
  adj2sg and sg2adj convert back and fort. 
  
  The edges can be transposed with method \code{t}.
  
  See also \code{sg2sym}
}
\details{
  \tabular{ll}{
    Date: \tab 2008-07-30\cr
    License: \tab GPL v2 or later\cr
  }
}
\keyword{spatial}