\name{spatgraphs-sg}
\alias{sg}
\alias{sgadj}
\alias{verifyclass}
\title{sg}

\description{
  Edge list-of-lists class for spatgraphs. Methods: print, plot.
  
  Older version used adjacency matrix, this class is now called sgadj.
  
  tosg and tosgadj convert back and fort.
  
  See also \code{sg2sym}
}
\details{
  \tabular{ll}{
    Date: \tab 2008-07-30\cr
    License: \tab GPL v2 or later\cr
  }
}
\keyword{spatial}